% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2_world}
\alias{co2_world}
\title{World CO2 emissions}
\format{
A data frame with 495 observations and 11 variables:
\itemize{
\item \code{iso_code} : ISO code A3 for country.
\item \code{name} : Name of country.
\item \code{co2} : Annual CO2 emissions - Annual total emissions of carbon dioxide (CO2), excluding land-use change, measured in million tonnes.
\item \code{co2_per_capita} : Annual CO2 emissions (per capita) - Annual total emissions of carbon dioxide (CO2), excluding land-use change, measured in tonnes per person.
\item \code{geometry} : Geographical attributes.
}
}
\source{
\href{https://github.com/owid/co2-data}{Our World In Data}
}
\usage{
co2_world
}
\description{
This dataset contains world polygons with CO2 emissions.
}
\note{
Documentation is from Our World In Data, see \url{https://github.com/owid/co2-data} for the data
and \url{https://ourworldindata.org/co2-and-greenhouse-gas-emissions} for more about CO2 emissions.
}
\keyword{datasets}
