% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries_gdp}
\alias{countries_gdp}
\title{Countries GDP}
\format{
A data frame with 177 observations and 3 variables:
\itemize{
\item \code{REGION_UN} : Continent
\item \code{SUBREGION} : Sub-region in the continent
\item \code{ADMIN} : Administrative name of country
\item \code{GDP_MD} : GDP
}
}
\source{
Package \href{https://github.com/ropensci/rnaturalearth}{rnaturalearth}
}
\usage{
countries_gdp
}
\description{
These data represent the GDP of the world's countries, classified by continent and sub-region.
This is a subset of the dataset \code{rnaturalearth::countries110}.
}
\keyword{datasets}
