% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3.R
\name{d3-format}
\alias{d3-format}
\alias{d3_format}
\alias{d3_format_time}
\title{Format numbers and dates with D3}
\usage{
d3_format(format, prefix = "", suffix = "", locale = "en-US")

d3_format_time(format, prefix = "", suffix = "", locale = "en-US")
}
\arguments{
\item{format}{Format for numbers, currency, percentage, e.g. \code{".0\\\%"} for rounded percentage.
See online documentation : \url{https://github.com/d3/d3-format}.}

\item{prefix}{Character string to append before formatted value.}

\item{suffix}{Character string to append after formatted value.}

\item{locale}{Localization to use, for example \code{"fr-FR"} for french,
see possible values here: \url{https://github.com/d3/d3-format/tree/master/locale}.}
}
\value{
a \code{JS} function.
}
\description{
Format numbers and dates with D3
}
\examples{

library(vchartr)
}
