% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco2mix}
\alias{eco2mix}
\title{Monthly electricity generation by source in France}
\format{
A data frame with 151 observations and 10 variables:
\itemize{
\item \code{date} : Date
\item \code{fuel} : Fuel generation in MW
\item \code{coal} : Coal generation in MW
\item \code{gas} : Gas generation in MW
\item \code{nuclear} : Nuclear generation in MW
\item \code{wind} : Wind generation in MW
\item \code{solar} : Solar generation in MW
\item \code{hydraulic} : Hydraulic generation in MW
\item \code{pumping} : Pumping generation in MW
\item \code{bioenergies} : Bioenergies generation in MW
}
}
\source{
\href{https://odre.opendatasoft.com/explore/dataset/eco2mix-national-cons-def/}{eco2mix}
}
\usage{
eco2mix
}
\description{
This dataset represents monthly electricity generation by source in France over the period 2012 - 2024.
}
\keyword{datasets}
