% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco2mix_long}
\alias{eco2mix_long}
\title{Monthly electricity generation by source in France (long format)}
\format{
A data frame with 1359 observations and 3 variables:
\itemize{
\item \code{date} : Date
\item \code{source} : Production according to the different sectors making up the energy mix.
\item \code{production} : Generation in MW
}
}
\source{
\href{https://odre.opendatasoft.com/explore/dataset/eco2mix-national-cons-def/}{eco2mix}
}
\usage{
eco2mix_long
}
\description{
This dataset represents monthly electricity generation by source in France over the period 2012 - 2024.
}
\keyword{datasets}
