% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electricity_mix}
\alias{electricity_mix}
\title{Electricity mix for 10 countries}
\format{
A data frame with 70 observations and 3 variables:
\itemize{
\item \code{country} : Country name
\item \code{source} : source of electricity
\item \code{generation} : Total electricity generation - Measured in terawatt-hours.
\item \code{type} : Low carbon or fossil fuels type of source.
}
}
\source{
\href{https://github.com/owid/energy-data}{Our World In Data}
}
\usage{
electricity_mix
}
\description{
This dataset represents the electricity mix of 10 countries (those with the highest electricity generation) in 2023.
}
\keyword{datasets}
