% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_sankey}
\alias{energy_sankey}
\title{Data for Sankey Chart}
\format{
A data frame with 177 observations and 3 variables:
\itemize{
\item \code{source} : Source
\item \code{target} : Target
\item \code{value} : Energy in TWh
}
}
\source{
Department of Energy & Climate Change via Tom Counsell
}
\usage{
energy_sankey
}
\description{
These data represent how energy is converted or transmitted before being consumed or lost.
}
\keyword{datasets}
