% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{JS}
\alias{JS}
\title{Mark character strings as literal JavaScript code}
\usage{
JS(...)
}
\arguments{
\item{...}{character vectors as the JavaScript source code (all arguments
will be pasted into one character string)}
}
\value{
A string that will be interpreted as JavaScript code in htmlwidgets.
}
\description{
See \code{\link[htmlwidgets:JS]{htmlwidgets::JS()}} for details.
}
\keyword{internal}
