% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{label_value}
\alias{label_value}
\title{Useful labeller functions}
\usage{
label_value(labels, multi_line = TRUE)
}
\arguments{
\item{labels}{Data frame of labels. Usually contains only one
element, but faceting over multiple factors entails multiple
label variables.}

\item{multi_line}{Whether to display the labels of multiple factors
on separate lines.}
}
\value{
Labels to be used in a facetted chart.
}
\description{
See \code{\link[ggplot2:labellers]{ggplot2::label_value()}} for details.
}
\keyword{internal}
