% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo_paris}
\alias{meteo_paris}
\title{Paris climate}
\format{
A data frame with 177 observations and 3 variables:
\itemize{
\item \code{month} : Month of the year
\item \code{temperature_avg} : Average temperature (°C)
\item \code{temperature_min} : Average minimum temperature (°C)
\item \code{temperature_max} : Average maximum temperature (°C)
\item \code{precipitation} : Precipitation (mm)
\item \code{humidity} : Humidity (\%)
\item \code{rainy_days} : Rainy days (days)
\item \code{sunshine_hours} : Sunshine hours (h)
}
}
\source{
\href{https://fr.climate-data.org/europe/france/ile-de-france/paris-44/#climate-table}{Climate-data.org}
}
\usage{
meteo_paris
}
\description{
This data contains information about the climate in Paris, France.
}
\keyword{datasets}
