% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale-discrete}
\alias{scale-discrete}
\alias{v_scale_x_discrete}
\alias{v_scale_y_discrete}
\title{Axis scale for discrete data}
\usage{
v_scale_x_discrete(vc, name = NULL, ..., position = "bottom")

v_scale_y_discrete(vc, name = NULL, ..., position = "left")
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}} or specific chart's type function.}

\item{name}{Title for the axis.}

\item{...}{Additional parameters for the axis.}

\item{position}{Position of the axis.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Axis scale for discrete data
}
\examples{

library(vchartr)

}
