% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{top_cran_downloads}
\alias{top_cran_downloads}
\title{Top CRAN downloads}
\format{
A data frame with 100 observations and 5 variables.
}
\source{
\href{https://r-hub.github.io/cranlogs/}{cranlogs}
}
\usage{
top_cran_downloads
}
\description{
The dataset contains data about CRAN downloads retrieved with \href{https://r-hub.github.io/cranlogs/}{cranlogs::cran_top_downloads}.
}
\keyword{datasets}
