% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{top_generation}
\alias{top_generation}
\title{Top electricity-generating countries}
\format{
A data frame with 10 observations and 2 variables:
\itemize{
\item \code{country} : Country name
\item \code{electricity_generation} : Total electricity generation - Measured in terawatt-hours.
}
}
\source{
\href{https://github.com/owid/energy-data}{Our World In Data}
}
\usage{
top_generation
}
\description{
This dataset represents the 10 countries with the highest electricity generation in 2023.
}
\keyword{datasets}
