% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{v_area}
\alias{v_area}
\title{Create an Area Chart}
\usage{
v_area(
  vc,
  mapping = NULL,
  data = NULL,
  name = NULL,
  stack = FALSE,
  curve_type = "linear",
  fill = NULL,
  fill_opacity = NULL,
  point = list(visible = FALSE),
  line = list(visible = FALSE),
  dataserie_id = NULL
)
}
\arguments{
\item{vc}{A chart initialized with \code{\link[=vchart]{vchart()}}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{data}{Default dataset to use for chart. If not already
a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{name}{Name for the serie, only used for single serie (no \code{color}/\code{fill} aesthetic supplied).}

\item{curve_type}{Curve interpolation type, see \href{https://www.visactor.io/vchart/option/AreaChart#area.style.curveType}{online documentation}.}

\item{fill}{Fill color.}

\item{fill_opacity}{Stroke color.}

\item{point}{Options for showing points on lines or not.}

\item{line}{Options for showing lines or not.}

\item{dataserie_id}{ID for the serie, can be used to customize the serie with \code{\link[=v_specs]{v_specs()}}.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create an Area Chart
}
\examples{

library(vchartr)

# Basic Area Chart
vchart(eco2mix) \%>\% 
  v_area(aes(date, solar), points = TRUE)

# Two areas
vchart(eco2mix, aes(date)) \%>\% 
  v_area(aes(y = solar)) \%>\% 
  v_area(aes(y = wind))

# Line chart with discrete x axis
vchart(data.frame(month = month.abb, value = sample(1:50, 12))) \%>\% 
  v_area(aes(month, value))

# Fill color 
vchart(data.frame(month = month.abb, value = sample(1:50, 12))) \%>\% 
  v_area(aes(month, value), fill = "firebrick", fill_opacity = 0.9)

# Smooth Area Chart
vchart(data.frame(month = month.abb, value = sample(1:50, 12))) \%>\% 
  v_area(aes(month, value), curve_type = "monotone")

# Step Area Chart
vchart(data.frame(month = month.abb, value = sample(1:50, 12))) \%>\% 
  v_area(aes(month, value), curve_type = "stepAfter")

# Multiple areas
vchart(eco2mix_long) \%>\% 
  v_area(aes(date, production, fill = source))

vchart(eco2mix_long) \%>\% 
  v_area(aes(date, production, fill = source), stack = TRUE, fill_opacity = 1)
}
