% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{v_labs}
\alias{v_labs}
\title{Set chart title and subtitle}
\usage{
v_labs(vc, title, subtitle = NULL)
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}} or specific chart's type function.}

\item{title}{Title for the chart.}

\item{subtitle}{Subtitle for the chart.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Set chart title and subtitle
}
\examples{
library(vchartr)
data("mpg", package = "ggplot2")

vbar(table(Class = mpg$class), aes(Class, Freq)) \%>\%
  v_labs(
    title = "Title for the chart",
    subtitle = "A subtitle to be placed under the title"
  )
}
