% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{v_line}
\alias{v_line}
\title{Create a Line Chart}
\usage{
v_line(
  vc,
  mapping = NULL,
  data = NULL,
  name = NULL,
  dataserie_id = NULL,
  points = FALSE,
  curve_type = "linear",
  line_dash = 0,
  stroke = NULL
)
}
\arguments{
\item{vc}{A chart initialized with \code{\link[=vchart]{vchart()}}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{data}{Default dataset to use for chart. If not already
a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{name}{Name for the serie, only used for single serie (no \code{color}/\code{fill} aesthetic supplied).}

\item{dataserie_id}{ID for the serie, can be used to customize the serie with \code{\link[=v_specs]{v_specs()}}.}

\item{points}{Show points on lines or not.}

\item{curve_type}{Curve interpolation type, see \href{https://www.visactor.io/vchart/option/lineChart#line.style.curveType}{https://www.visactor.io}.}

\item{line_dash}{Used to configure the dashed line mode when filling lines.
It uses an array of values to specify the alternating lengths of lines and gaps that describe the pattern.}

\item{stroke}{Stroke color.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create a Line Chart
}
\examples{

library(vchartr)

# Basic Line Chart
vchart(eco2mix) \%>\% 
  v_line(aes(date, solar))

# Line chart with discrete x axis
vchart(data.frame(month = month.abb, value = sample(1:50, 12))) \%>\% 
  v_line(aes(month, value))

# Stroke color 
vchart(data.frame(month = month.abb, value = sample(1:50, 12))) \%>\% 
  v_line(aes(month, value), stroke = "firebrick")

# Smooth Line Chart
vchart(data.frame(month = month.abb, value = sample(1:50, 12))) \%>\% 
  v_line(aes(month, value), curve_type = "monotone")

# Step Line Chart
vchart(data.frame(month = month.abb, value = sample(1:50, 12))) \%>\% 
  v_line(aes(month, value), curve_type = "stepAfter")

# Dash array 
vchart(data.frame(month = month.abb, value = sample(1:50, 12))) \%>\% 
  v_line(aes(month, value), line_dash = c(10, 10))

# Multiple lines
vchart(eco2mix_long) \%>\% 
  v_line(aes(date, production, color = source))

}
