% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{v_line_datazoom}
\alias{v_line_datazoom}
\title{Add data zoom on a line or area chart}
\usage{
v_line_datazoom(
  vc,
  start = "{label:\%Y-\%m-\%d}",
  end = "{label:\%Y-\%m-\%d}",
  brush = TRUE
)
}
\arguments{
\item{vc}{A chart created with \code{\link[=vchart]{vchart()}}.}

\item{start, end}{Formatter for the start/end label, e.g. : \code{"Start: \\{label:\%Y-\%m-\%d\\}"},
where the part between braces will be replaced by the date with the format specified.}

\item{brush}{Logical, add the ability to brush the chart to zoom in.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Add data zoom on a line or area chart
}
\examples{
library(vchartr)
data("economics", package = "ggplot2")
vchart(economics, aes(date, unemploy)) \%>\%
  v_line() \%>\%
  v_line_datazoom()
}
