% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{v_specs}
\alias{v_specs}
\title{Specify configuration options for a \code{\link[=vchart]{vchart()}}.}
\usage{
v_specs(vc, ...)
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}} or specific chart's type function.}

\item{...}{List of options to specify for the chart, see \url{https://www.visactor.io/vchart/option/}.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Specify configuration options for a \code{\link[=vchart]{vchart()}}.
}
\examples{
library(vchartr)
data("mpg", package = "ggplot2")

vbar(table(Class = mpg$class), aes(Class, Freq)) \%>\%
  v_specs(
    label = list(visible = TRUE),
    color = list("firebrick")
  )
}
