% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{v_specs_colors}
\alias{v_specs_colors}
\title{Set color(s) for chart}
\usage{
v_specs_colors(vc, ...)
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}}.}

\item{...}{Colors options, can be a single color code, a vector of colors to use or a list with more options.
For \code{v_colors_manual} it should be a named list with data values as name and color as values.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Set color(s) for chart
}
\examples{
library(vchartr)
data("mpg", package = "ggplot2")

vchart(table(Class = mpg$class)) \%>\%
  v_bar(aes(Class, Freq)) \%>\%
  v_specs_colors("#8FBCBB")

}
