% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{v_theme_builtin}
\alias{v_theme_builtin}
\title{Builtin themes for Charts}
\usage{
v_theme_builtin(vc, name)
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}}.}

\item{name}{Name of the theme to use, see \href{https://www.visactor.io/vchart/guide/tutorial_docs/Theme/Theme_Extension}{available themes online}.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Builtin themes for Charts
}
\examples{

library(vchartr)

vchart(top_generation) \%>\%
  v_bar(aes(country, electricity_generation)) \%>\%
  v_theme_builtin("dark")

vchart(top_generation) \%>\%
  v_bar(aes(country, electricity_generation)) \%>\%
  v_theme_builtin("vScreenVolcanoBlue")

vchart(top_generation) \%>\%
  v_bar(aes(country, electricity_generation)) \%>\%
  v_theme_builtin("vScreenClean")

vchart(top_generation) \%>\%
  v_bar(aes(country, electricity_generation)) \%>\%
  v_theme_builtin("chartHubLight")

}
