% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart.R
\name{vchart}
\alias{vchart}
\title{Create interactive charts with VChart}
\usage{
vchart(data, mapping, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Can be a \code{data.frame} if function used with other layers functions or a list of parameters for configuring a chart.}

\item{mapping}{Default list of aesthetic mappings to use for chart, only used if \code{data} is a \code{data.frame}.}

\item{...}{Additional parameters.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
VChart is a charting component library, see more about it here : \url{https://www.visactor.io/vchart}.
}
\note{
This function allow you to use JavaScript function \code{VChart} directly,
see \url{https://www.visactor.io/vchart/option/} for how to specify options.
}
\examples{

library(vchartr)

vchart(
  type = "bar",
  data = list(
    list(
      id = "bardata",
      values = list(
        list(day = "Monday", value = 12),
        list(day = "Tuesday", value = 8),
        list(day = "Wednesday", value = 18),
        list(day = "Thursday", value = 25),
        list(day = "Friday", value = 14),
        list(day = "Saturday", value = 5),
        list(day = "Sunday", value = 7)
      )
    )
  ),
  xField = "day",
  yField = "value",
  crosshair = list(
    xField = list(visible = TRUE)
  )
)
}
