% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-package.R
\name{vchartr-exports}
\alias{vchartr-exports}
\alias{\%>\%}
\alias{aes}
\alias{vars}
\alias{label_value}
\alias{JS}
\title{vchartr exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the vchartr package to avoid listing the magrittr
as Depends of vchartr
}
