% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-package.R
\docType{package}
\name{vchartr-package}
\alias{vchartr}
\alias{vchartr-package}
\title{An \code{htmlwidget} interface to the
VChart javascript chart library}
\description{
This package allow you to use VChart.js (\url{https://www.visactor.io/vchart}),
to create interactive charts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dreamrs.github.io/vchartr/}
}

}
\author{
Victor Perrier (@dreamRs_fr)
}
