% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-types.R
\name{vcirclepacking}
\alias{vcirclepacking}
\title{Create a Circle Packing Chart}
\usage{
vcirclepacking(
  data,
  mapping,
  drill = TRUE,
  use_root = FALSE,
  fill_opacity = JS("d => d.isLeaf ? 0.75 : 0.25;"),
  label_visible = JS("d => d.depth === 1;"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Default dataset to use for chart. If not already
a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{drill}{Drill-down function switch.}

\item{use_root}{Add a root level in the hierarchy, can be \code{TRUE}
(in this case root level will be named \code{root}) or a \code{character} (use as the name for the root level).}

\item{fill_opacity}{Fill opacity, a JS function determining the opacity of the elements.}

\item{label_visible}{A JS function to control visibility of labels.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create a Circle Packing Chart
}
\examples{

library(vchartr)

vcirclepacking(
  countries_gdp,
  aes(lvl1 = REGION_UN, lvl2 = SUBREGION, lvl3 = ADMIN, value = GDP_MD)
)

# With root level
vcirclepacking(
  countries_gdp,
  aes(lvl1 = REGION_UN, lvl2 = SUBREGION, lvl3 = ADMIN, value = GDP_MD),
  use_root = "World"
)

# Custom colors
vcirclepacking(
  countries_gdp,
  aes(lvl1 = REGION_UN, lvl2 = SUBREGION, lvl3 = ADMIN, value = GDP_MD)
) \%>\%
  v_colors_manual(
    Oceania = "#E6AB02",
    Africa = "#1B9E77",
    Americas = "#D95F02",
    Asia = "#E7298A",
    Europe = "#66A61E",
    Antarctica = "#7570B3"
  )


# Bubble Chart
vcirclepacking(
  countries_gdp,
  aes(lvl = ADMIN, value = GDP_MD),
  label_visible = JS("d => d.value > 261921;"), # 261921 = 3rd Qu.
)
}
