% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-types.R
\name{vheatmap}
\alias{vheatmap}
\title{Create an Heatmap Chart}
\usage{
vheatmap(
  data,
  mapping,
  dataserie_id = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Can be a \code{data.frame} if function used with other layers functions or a list of parameters for configuring a chart.}

\item{mapping}{Default list of aesthetic mappings to use for chart, only used if \code{data} is a \code{data.frame}.}

\item{dataserie_id}{ID for the serie, can be used to customize the serie with \code{\link[=v_specs]{v_specs()}}.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create an Heatmap Chart
}
\examples{

library(vchartr)

# Heatmap with continuous fill variable
vheatmap(
  co2_emissions,
  aes(x = year, y = country, fill = co2_per_capita)
)

vheatmap(
  co2_emissions,
  aes(x = year, y = country, fill = co2_per_capita)
) \%>\%
  v_specs_colors(
    range = rev(
      c("#8C510A", "#BF812D", "#DFC27D", "#F6E8C3",
        "#C7EAE5", "#80CDC1", "#35978F", "#01665E")
    )
  )



# Heatmap with discrete fill variable
vheatmap(
  co2_emissions,
  aes(x = year, y = country, fill = co2_growth_change)
)

vheatmap(
  co2_emissions,
  aes(x = year, y = country, fill = co2_growth_change)
) \%>\%
  v_specs_colors_manual(
    Increase = "firebrick",
    Decrease = "forestgreen"
  )

}
