% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-types.R
\name{vhist}
\alias{vhist}
\title{Create an histogram}
\usage{
vhist(
  data,
  mapping,
  bins = 30,
  binwidth = NULL,
  stack = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Default dataset to use for chart. If not already
a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30.}

\item{binwidth}{The width of the bins. Can be specified as a numeric value
or as a function that calculates width from unscaled x. Here, "unscaled x"
refers to the original x values in the data, before application of any
scale transformation. When specifying a function along with a grouping
structure, the function will be called once per group.
The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.

The bin width of a date variable is the number of days in each time; the
bin width of a time variable is the number of seconds.}

\item{stack}{Whether to stack the data or not (if \code{fill} aesthetic is provided).}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create an histogram
}
\examples{

library(vchartr)
library(palmerpenguins)

vhist(penguins, aes(flipper_length_mm))

vhist(penguins, aes(flipper_length_mm)) \%>\%
  v_specs(
    bar = list(
      style = list(
        stroke = "white",
        lineWidth = 1
      )
    )
  )

vhist(penguins, aes(flipper_length_mm, fill = species))
vhist(penguins, aes(flipper_length_mm, fill = species), stack = TRUE)

vhist(penguins, aes(flipper_length_mm, fill = species)) \%>\%
  v_colors("#ffa232", "#b34df2", "#33a2a2") \%>\%
  v_specs(
    bar = list(
      style = list(opacity = 0.5)
    )
  )
}
