% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-types.R
\name{vline-varea}
\alias{vline-varea}
\alias{vline}
\alias{varea}
\alias{v_add_line}
\alias{v_add_range_area}
\title{Create a Line Chart}
\usage{
vline(
  data,
  mapping,
  format_date = "\%Y-\%m-\%d",
  format_datetime = "\%Y-\%m-\%d \%H:\%M",
  name = NULL,
  dataserie_id = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)

varea(
  data,
  mapping,
  format_date = "\%Y-\%m-\%d",
  format_datetime = "\%Y-\%m-\%d \%H:\%M",
  name = NULL,
  dataserie_id = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)

v_add_line(vc, mapping, data = NULL, name = NULL, dataserie_id = NULL)

v_add_range_area(vc, mapping, data = NULL, name = NULL, dataserie_id = NULL)
}
\arguments{
\item{data}{Default dataset to use for chart. If not already
a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{format_date}{Format to be applied if \code{x} aesthetic is a \code{Date}.}

\item{format_datetime}{Format to be applied if \code{x} aesthetic is a \code{POSIXt}.}

\item{name}{Name for the serie, only used for single serie (no \code{color} aesthetic supplied).}

\item{dataserie_id}{ID for the serie, can be used to customize the serie with \code{\link[=v_specs]{v_specs()}}.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}

\item{vc}{A chart created with \code{\link[=vline]{vline()}}.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create a Line Chart
}
\examples{

library(vchartr)

# Basic Line Chart
vline(eco2mix, aes(date, solar))

# Basic Area Chart
varea(eco2mix, aes(date, solar))

# Smooth Line Chart
vline(tail(eco2mix, 10), aes(date, solar), dataserie_id = "line") \%>\%
  v_specs(
    line = list(
      style = list(curveType = "monotone")
    ),
    dataserie_id = "line"
  )

# Step Line Chart
vline(tail(eco2mix, 10), aes(date, solar), dataserie_id = "line") \%>\%
  v_specs(
    line = list(
      style = list(curveType = "stepAfter")
    ),
    dataserie_id = "line"
  )


# Add Line to existing chart
vline(eco2mix, aes(date, solar)) \%>\%
  v_add_line(aes(date, wind))

# Use long format for multiple lines
vline(eco2mix_long, aes(date, production, color = source))
varea(eco2mix_long, aes(date, production, fill = source))

# Add a range area
vline(temperatures, aes(date, `2024`)) \%>\%
  v_add_line(aes(date, average)) \%>\%
  v_add_range_area(aes(date, ymin = low, ymax = high)) \%>\%
  v_legend(visible = TRUE)

# Change opacity & color for the area range
vline(temperatures, aes(date, `2024`)) \%>\%
  v_add_line(aes(date, average)) \%>\%
  v_add_range_area(
    aes(date, ymin = low, ymax = high),
    dataserie_id = "range_area"
  ) \%>\%
  v_legend(visible = TRUE) \%>\%
  v_specs(
    area = list(
      style = list(
        fillOpacity = 1,
        fill = "#D8DEE9"
      )
    ),
    dataserie_id = "range_area"
  )

}
