% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-types.R
\name{vline}
\alias{vline}
\title{Create a line chart}
\usage{
vline(
  data,
  mapping,
  format_date = "\%Y-\%m-\%d",
  format_datetime = "\%Y-\%m-\%d \%H:\%M",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Default dataset to use for chart. If not already
a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{format_date}{Format to be applied if \code{x} aesthetic is a \code{Date}.}

\item{format_datetime}{Format to be applied if \code{x} aesthetic is a \code{POSIXt}.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create a line chart
}
