% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-types.R
\name{vtreemap}
\alias{vtreemap}
\title{Create a Treemap Chart}
\usage{
vtreemap(
  data,
  mapping,
  drill = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Can be a \code{data.frame} if function used with other layers functions or a list of parameters for configuring a chart.}

\item{mapping}{Default list of aesthetic mappings to use for chart, only used if \code{data} is a \code{data.frame}.}

\item{drill}{Drill-down function switch.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create a Treemap Chart
}
\examples{

library(vchartr)

# Basic Treemap Chart
vtreemap(
  countries_gdp,
  aes(lvl1 = REGION_UN, lvl2 = ADMIN, value = GDP_MD)
) \%>\% 
  v_specs(
    label = list(visible = TRUE)
  )

# Show all levels
vtreemap(
  countries_gdp,
  aes(lvl1 = REGION_UN, lvl2 = SUBREGION, lvl3 = ADMIN, value = GDP_MD)
) \%>\% 
  v_specs(
    label = list(visible = TRUE),
    nonLeaf = list(visible = TRUE),
    nonLeafLabel = list(visible = TRUE, position = "top")
  )
}
