% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-types.R
\name{vwordcloud}
\alias{vwordcloud}
\title{Create a Wordcloud}
\usage{
vwordcloud(data, mapping, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Default dataset to use for chart. If not already
a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create a Wordcloud
}
\examples{

if (rlang::is_installed("cranlogs")) {
  
  top_pkgs <- cranlogs::cran_top_downloads(count = 50)
  
  vwordcloud(top_pkgs, aes(word = package, count = count))
  
  vwordcloud(top_pkgs, aes(word = package, count = count, color = package))
  
  vwordcloud(top_pkgs, aes(word = package, count = count)) \%>\% 
    v_specs(
      wordCloudConfig = list(
        zoomToFit = list(
          enlarge = TRUE,
          fontSizeLimitMax = 30
        )
      )
    )
  
}

}
