% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_electricity}
\alias{world_electricity}
\title{World low carbon & fossil electricity generation 2014 - 2023}
\format{
A data frame with 70 observations and 4 variables:
\itemize{
\item \code{year} : Year
\item \code{source} : Either :
\itemize{
\item \verb{Low carbon} : Electricity generation from low-carbon sources -
Low-carbon sources correspond to renewables and nuclear power, that produce significantly less greenhouse-gas emissions than fossil fuels.
\item \code{Renewables} : Electricity generation from renewables
\item \code{Nuclear} : Electricity generation from nuclear
\item \code{Fossil} : Electricity generation from fossil fuels (oil + gas + coal)
\item \code{Oil} : Electricity generation from fossil fuels
\item \code{Gas} : Electricity generation from fossil fuels
\item \code{Coal} : Electricity generation from fossil fuels
}
\item \code{generation} : Electricity generation in terawatt-hours.
\item \code{type} : Type of source : total or detail.
}
}
\source{
\href{https://github.com/owid/energy-data}{Our World In Data}
}
\usage{
world_electricity
}
\description{
This dataset represents world's electricity generation from low-carbon sources and fossil fuels over the period 2014 - 2023.
}
\keyword{datasets}
