% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_generation}
\alias{world_generation}
\title{World low carbon & fossil generation 2014 - 2023}
\format{
A data frame with 20 observations and 3 variables:
\itemize{
\item \code{year} : Year
\item \code{source} : Either :
\itemize{
\item \verb{Low carbon} : Electricity generation from low-carbon sources - Low-carbon sources correspond to renewables and nuclear power, that produce significantly less greenhouse-gas emissions than fossil fuels.
\item \code{Fossil} : Electricity generation from fossil fuels
}
\item \code{generation} : Electricity generation in terawatt-hours.
}
}
\source{
\href{https://github.com/owid/energy-data}{Our World In Data}
}
\usage{
world_generation
}
\description{
This dataset represents world's electricity generation from low-carbon sources and fossil fuels over the period 2014 - 2023.
}
\keyword{datasets}
