% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale-continuous}
\alias{scale-continuous}
\alias{v_scale_x_continuous}
\alias{v_scale_y_continuous}
\title{Axis scale for continuous data}
\usage{
v_scale_x_continuous(
  vc,
  name = NULL,
  breaks = NULL,
  pretty = TRUE,
  labels = NULL,
  zero = FALSE,
  min = NULL,
  max = NULL,
  position = "bottom"
)

v_scale_y_continuous(
  vc,
  name = NULL,
  breaks = NULL,
  pretty = TRUE,
  labels = NULL,
  zero = FALSE,
  min = NULL,
  max = NULL,
  position = "left"
)
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}} or specific chart's type function.}

\item{name}{Title for the axis.}

\item{breaks}{One of:
\itemize{
\item A single \code{numeric} value giving the number of breaks.
\item A numeric vector of positions.
}}

\item{pretty}{Use \code{\link[=pretty]{pretty()}} to dertimen breaks if \code{breaks} is a single numeric value.}

\item{labels}{The format to be applied on numeric in the labels. Either:
\itemize{
\item A single character indicationg th D3 format.
\item A \code{JS} function, such as \code{\link[=d3_format]{d3_format()}}.
}}

\item{zero}{Force axis to start at 0.}

\item{min}{Minimum value on the axis.}

\item{max}{Maximum value on the axis.}

\item{position}{Position of the axis.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Axis scale for continuous data
}
\examples{

library(vchartr)

vbar(
  top_generation,
  aes(country, electricity_generation)
) \%>\%
  v_scale_y_continuous(name = "Electricity generation")

vbar(
  top_generation,
  aes(country, electricity_generation)
) \%>\%
  v_scale_y_continuous(breaks = 10)


vbar(
  top_generation,
  aes(country, electricity_generation)
) \%>\%
  v_scale_y_continuous(breaks = c(0, 5000, 10000))


vbar(
  top_generation,
  aes(country, electricity_generation)
) \%>\%
  v_scale_y_continuous(labels = "~s")


vbar(
  top_generation,
  aes(country, electricity_generation)
) \%>\%
  v_scale_y_continuous(labels = d3_format(",", suffix = " TWh", locale = "fr-FR"))

}
