% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale-date}
\alias{scale-date}
\alias{v_scale_x_date}
\alias{v_scale_y_date}
\title{Axis scale for date/time data}
\usage{
v_scale_x_date(
  vc,
  name = NULL,
  date_breaks = NULL,
  date_labels = NULL,
  min = NULL,
  max = NULL,
  position = "bottom"
)

v_scale_y_date(
  vc,
  name = NULL,
  date_breaks = NULL,
  date_labels = NULL,
  min = NULL,
  max = NULL,
  position = "left"
)
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}} or specific chart's type function.}

\item{name}{Title for the axis.}

\item{date_breaks}{One of:
\itemize{
\item A single \code{numeric} value giving the number of breaks.
\item A string giving the distance between breaks like "2 weeks", or "10 years".
\item A Date/POSIXct vector giving positions of breaks.
}}

\item{date_labels}{The format to be applied on Date/POSIXct in the labels.}

\item{min}{Minimum value on the axis.}

\item{max}{Maximum value on the axis.}

\item{position}{Position of the axis.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Axis scale for date/time data
}
\examples{

library(vchartr)

vline(eco2mix, aes(as.numeric(date), solar)) \%>\%
  v_scale_x_date(name = "Date")


vline(eco2mix, aes(as.numeric(date), solar)) \%>\%
  v_scale_x_date(date_breaks = 10)

vline(eco2mix, aes(as.numeric(date), solar)) \%>\%
  v_scale_x_date(date_breaks = "2 years")

}
