% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{v_specs_indicator}
\alias{v_specs_indicator}
\title{Add indicator to chart}
\usage{
v_specs_indicator(vc, ...)
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}}.}

\item{...}{Options for the legend, see examples or
\href{https://www.visactor.io/vchart/option/commonChart#indicator}{online documentation}.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Add indicator to chart
}
