% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{v_specs_player}
\alias{v_specs_player}
\title{Set player options}
\usage{
v_specs_player(vc, ...)
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}} or specific chart's type function.}

\item{...}{Options for the legend, see examples or
\href{https://www.visactor.io/vchart/option/commonChart#player}{online documentation}.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Set player options
}
\examples{
library(vchartr)
data("mpg", package = "ggplot2")

vchart(table(Class = mpg$class, Year = mpg$year)) \%>\%
  v_bar(aes(Class, Freq, fill = Year)) \%>\%
  v_specs_tooltip(
    visible = FALSE
  )
}
