% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{v_specs_tooltip}
\alias{v_specs_tooltip}
\title{Set tooltip options}
\usage{
v_specs_tooltip(vc, ...)
}
\arguments{
\item{vc}{An htmlwidget created with \code{\link[=vchart]{vchart()}} or specific chart's type function.}

\item{...}{Options for the legend, see examples or online documentation :
\url{https://www.visactor.io/vchart/guide/tutorial_docs/Chart_Concepts/Tooltip}.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Set tooltip options
}
\examples{
library(vchartr)
data("mpg", package = "ggplot2")

vbar(
  table(Class = mpg$class, Year = mpg$year),
  aes(Class, Freq, fill = Year)
) \%>\%
  v_specs_tooltip(
    visible = FALSE
  )
}
