% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-types.R
\name{vbar}
\alias{vbar}
\title{Create a bar chart}
\usage{
vbar(
  data,
  mapping,
  stack = FALSE,
  percent = FALSE,
  direction = c("vertical", "horizontal"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Default dataset to use for chart. If not already
a \code{data.frame}, it will be coerced to with \code{as.data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{stack}{Whether to stack the data.}

\item{percent}{Whether to display the data as a percentage.}

\item{direction}{The direction configuration of the chart: \code{"vertical"} (default) or \code{"horizontal"}.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create a bar chart
}
\examples{

library(vchartr)

# Classic Bar Chart
vbar(
  top_generation,
  aes(country, electricity_generation)
)

# Horizontal Bar Chart
vbar(
  top_generation,
  aes(country, electricity_generation),
  direction = "horizontal"
)


# Grouped Bar Chart
vbar(
  world_generation,
  aes(year, generation, fill = source)
)

# Horizontal Grouped Bar Chart
vbar(
  world_generation,
  aes(year, generation, fill = source),
  direction = "horizontal"
)

# Stacked Bar Chart
vbar(
  world_generation,
  aes(year, generation, fill = source),
  stack = TRUE
)

# Percentage Stacked Bar Chart
vbar(
  world_generation,
  aes(year, generation, fill = source),
  stack = TRUE,
  percent = TRUE
)

}
