% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vchart-types.R
\name{vscatter}
\alias{vscatter}
\title{Create a Scatter Chart}
\usage{
vscatter(
  data,
  mapping,
  dataserie_id = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Can be a \code{data.frame} if function used with other layers functions or a list of parameters for configuring a chart.}

\item{mapping}{Default list of aesthetic mappings to use for chart, only used if \code{data} is a \code{data.frame}.}

\item{dataserie_id}{ID for the serie, can be used to customize the serie with \code{\link[=v_specs]{v_specs()}}}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=vchart]{vchart()}} \code{htmlwidget} object.
}
\description{
Create a Scatter Chart
}
\examples{

library(vchartr)
library(palmerpenguins)

vscatter(penguins, aes(x = flipper_length_mm, y = body_mass_g))

vscatter(
  penguins,
  aes(x = flipper_length_mm, y = body_mass_g, color = species)
)

vscatter(
  penguins,
  aes(
    x = bill_length_mm,
    y = bill_depth_mm,
    color = species,
    shape = species
  ),
  dataserie_id = "penguins"
) \%>\%
  v_specs(
    shape = list(
      type = "ordinal",
      range = c("circle", "square", "triangle")
    ),
    dataserie_id = "penguins"
  )

vscatter(
  penguins,
  aes(x = flipper_length_mm, y = body_mass_g, color = species)
) \%>\%
  v_specs_colors_manual(
    Adelie = "#ffa232",
    Chinstrap = "#33a2a2",
    Gentoo = "#b34df2"
  )
}
